Removed cucumber features and dependency.
Cucumber really was not the right tool for the job.
I only chose it, because I was working with it on
another project at the time.

Also, machinist was really nice, but is still in
beta after all this time. So I switched to
factory_girl, which seems to be much more popular.