Reactivated integrated mailing functionality for accepting/ rejecting events.
The main extension to the old implementation is, that notifications
can be changed over the call for papers interface for every
conference independently.

The multilingual logic based on globalize3 gem. There is a
notification model for every call for papers with the possibility to
store a accept/reject body and subject for every conference
language. Every speaker is contacted in his first language added
during the account creation. The fall-back language is english.

There is also a change for the session_store. All session data is from
now on stored in a sessions table in the database.