create static export in tmp directory and offer download
While it was sure nice to create the static schedule export directly
below public/ and have access via the same domain it leaves the service
open for XSS via attachments. While this could, again, be mitigated by
configuring the http server in front of frab correcly, it's probably
easier to serve the schedule from another domain.