expenses: add model, views and controller
Add a model called Expense to store expenses associated with a Person
and a Conference. An expense stores a name, a value and whether or not
it has been reimbursed to the person.

Update the Person class so it has_many expenses, and nest the expenses
resources under the person in the routing. The controller is a
straight-forward implementation of a nested resource controller.

Introduce a new tab in the People view, and show all current expenses as
table, with associated action buttons. Allow updating and creating new
expenses via forms. Also sum up the value for all reimbursed and
non-reimbursed entries and display the numbers underneath the table view.