events: add action to copy an event from one conference to another
When an event can't happen in one year due to time limitiations, and it is
already created, it is very handy to have a way to simply make a duplicate
of the event and store it into another conference.

This is what this patch implements. The associated people and the
attachments are also duplicated, but the rest is left unset. The exising
(old) record is not altered in any way, the new record is a full clone of
the old one.