statistics: add gender breakdown
Join events of a conference with persons involved, and group the resulting
data set by gender. Use that to provide a json array which can in turn be
used by the jQuery plot module.

The breakdown is limited to moderators or speakers of an event, and to
events that are public.

Display a plot on the home screen of a conference.