conferences, events, people: switch to ransack for full-text search
Ransack is already included in the gems, so all we have to do is cut out the
acts_as_indexed pieces and use ransack directly in the models.

For this, the controllers for conferences, events and people have been
changed so that the ransack filters are called directly, using the 'cond'
predicate in most cases.

The search code in controllers was factored out into a new method called
'search'. Depending on whether 'q' or 'term' has been passed in through
the query, it uses the model's search method or invokes ransack directly.
This way, both query methods are still offered.