Add MailTemplates
A MailTemplate belongs_to a conference features a name, a subject and
a content.

The model support place holders such as "#first_name", "#last_name" and
"#public_name" that are replaced by a Person's data when sending the mail.

Add send form in mail_templates/show with a select box for pre-defined
filters to define which persons should receive the mails. These filters
are easy to extend and behave a bit like reports. If you need more, just
add what you need.

The actual send logic is implemented in a SuckerPunch job and is executed
asnychronously when running in the production environment. Otherwise, it is
ran synchronously for better debug possibilities.