event: calculate available start times in event form
In the form for events, intersect the availability for all presenting
persons, and only display those as options in the start_time drop down
menu.

In order not to break records that have a conflicting time set, the current
value is special-cased and marked accordingly. This way, people can still
save such an event without accidentally altering the start time.

For this to work, the SimpleForm wrapper is provided with a block, so the
default Rails form helper can kick in and make use of
grouped_options_for_select().