Overhaul ticket server implementation
Currently, when a ticker server is configured, the conference model
will pass back either the RTTickets or the OtrsTickets module through
get_ticket_module. The caller will then pass in the conference to the
functions in the module, through which the code will then access the
TicketServer instance.

This seems a little backwards, so this patch turns that around: The
TicketServer class will now dynamically instanciate either a
OTRSTicketServerAdapter or an RTTicketServerAdapter, which are both
subclasses of TicketServerAdapter. The TicketServerAdapter has access to the
TicketServer and derives all information directly from it.

This leads to some code de-duplication and a clearer object model.

Behavior is identical in at least superficial tests with local OTRS and RT
instances.