Add support for Redmine tickets
Allow users to configure a Redmine ticket server backend.
This implementation lets ActiveRessource handle all the json/REST magic.

The 'password' field of the configuration carries an API key instead, as
this is how their REST API works, so I tweaked the help text a bit.

From TicketsHelper::get_ticket_view_url, use Conference::get_ticket_module,
as that already does all the string-to-module dance for us.

The patch also adds an :event_url to the hashmap that is passed to the
create_remote_ticket implementations, so the modules can include a URL to
the ticket that points back to the frab event.