rake: make fake data generator more powerful
When a new frab instance is set up for development, it is very useful to
have some fake data to play around with. The current fake data generator
bails out if no conference was added by other means, which make it
cumbersome to populate an empty database.

Fix this by restructuring the code so it generates a bunch of new
conferences and nests new tracks, days, rooms, languages, events,
availabilities etc. under it.