conference: allow nesting under parent
Add a 'parent' field to conferences to conferences can be nested, and
allow accessing sub-conferences though a has_many :sub_conferences
assiciation.

When a conference is nested under a parent, it shares the days, timeslot
configuration and the timezone of the parent. Hence, override the
necessary methods and do transparent up-chaining.

Also, add a 'events_including_sub_conferences' helper to access all events
from both the parent and all its sub-conferences.