Use a random password when seeding the database in production mode
Default passwords suck, and administrators of production machines tend to
forget to change them after the initial setup.

Use a random password when seeding the database in production
environment, but stick to 'test123' for development mode. Also print
the chosen password to the console, so people can actually log in.