Make welcome controller responsible for dispatching to cfp pages
Since the welcome controller is not very useful once you're logged in, it now redirects to the appropriate page:

* the edit profile page, if public_name still is user.email
* the cfp user page that shows your events and allows adding or joining new events, if your proifle has been filled
* the conference management page for crew

The sign_in now redirects back to the ressource that triggered the sign in request or to the cfp root path which
then redirects further as outlined above.

The tests have been modified accordingly and now work.