Change 'enabled' string logic
The descriptive strings are currently only used in a view for the
conference settings, under the heading 'additional features' with
checkboxes beside them.

From a logical point of view, writing "enabled" and having a checkbox is
fairly redundant (if a checkbox is checked, something is active, after
all); thus I propose removing the "enabled" bit of the string and have
done this in all the languages whose characters I can decipher and know
if the grammar would be impacted by a change. (Spoilers: it isn't for
romance or germanic languages)

Additionally, some of the strings were incorrectly referenced as
'expenses enabled', without the underscore, which I've supplied.