import omniauth data
when exporting using frab:conference_export, the omniauth data (used to
enable logging in with 3rd party authenticators) is exported. With this
commit, then this data is also used when importing with
frab:conference_import); so a user which was able to log in to the
original frab with her Google account will now be able to log in to the
new frab as well.

Existing users on the destination frab are not modified.

This is a completion of https://github.com/frab/frab/issues/499