add assistant
A new user role in an event: "assistant" . Assistant has the same
permissions as a speaker, but their availability does not count.
The assistant is notified by mail of changes to the event, similar to
presenter.

Prospective assistants can join events by use of the join token,
similar to co-speakers.

The group of people who needs to be notified by e-mail when something
is changed with the event is now called "subscribers".

see frab#488