Remove migration test from CI
Previously we assumed that `db/schema.rb` is database agnostic.
This is not exactly true in all regards.

* https://github.com/rails/rails/issues/26209#issuecomment-243861875

Our commited `schema.rb` can be used to `db:setup` all databases.
running `db:schema:dump` or `db:migrate`, will turn the file into a
database specific one.
Formerly that wasn't obvious with the sqlite adapter, but now it dumps
foreign key information into the schema.
The previous check cannot be used anymore.

This is unfortunate, since frab's migrations are so old, that you cannot
`db:create db:migrate` a fresh system.

Not sure if related, but we will probably see more problems with old
frab installations on MySQL, due to the 2016 bigint change. These become
especially problematic, when using foreign key constraints:

* (MySQL switched to bigint)[https://github.com/rails/rails/pull/26266]
* https://github.com/rails/rails/issues/43168

Changing all IDs and references to bigint on MySQL is probably not
possible from migrations and should be done by the admin. A more
extensive solution than this is needed:
https://github.com/frab/frab/issues/660#issuecomment-557502669