Support lang param in public schedule
We use the `locale` param for interface translations. Conferences can
store field data, e.g. descriptions, in a different set of languages.

The `lang` param is used to choose the language of thosse fields.

The locale selection now also falls back to the I18n.default_locale and
is no longer hardcoded to 'en'.