Registration form includes person fields
The registration form was previously kept simple by design. However,
that meant the person's public_name had to be set to a temporary
value. Later code tried to redirect users to the person edit form, so
the could fix up their public name. That didn't work all the time and
resulted in unexpected behavior.